Rename request.rs -> request_ext.rs
