+++ src/lib.rs
@@ -525,7 +525,7 @@ use http::{Request, Response, Uri};
 pub use proxy::Proxy;
 pub use request::RequestBuilder;
 use request::{WithBody, WithoutBody};
-pub use response::ResponseExt;
+pub use response_ext::ResponseExt;
 pub use send_body::AsSendBody;
 
 mod agent;
@@ -536,7 +536,7 @@ mod pool;
 mod proxy;
 mod query;
 mod request;
-mod response;
+mod response_ext;
 mod run;
 mod send_body;
 mod timings;
